% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Draw accidents on a map by specified state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A character or number which represents a state.}

\item{year}{A character or number which represents a year.}
}
\value{
a chart wich show accidents on a map.
}
\description{
This function reads archive by specified year from working directory and
draw accidents on a map by specified state and year
}
\examples{
\dontrun{
fars_map_state(4,2013)
fars_map_state("6",2013)
}

}
