% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a delimited file into a tibble}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving a path to a file.}
}
\value{
The function returns a data frame.
}
\description{
This function reads comma separated data from specified file path.
Files ending in .gz, .bz2, .xz, or .zip containing a csv file will be
automatically uncompressed.
Files should be stored locally. Remote files can not be automatically downloaded.
}
\examples{
fars_read(file.path("data", "accident_2013.csv.bz2"))
fars_read("accident_2013.csv")

\dontrun{
fars_read("https://github.com/tidyverse/readr/raw/master/inst/extdata/mtcars.csv")
}

}
