% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read files by years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A character or numeric vector which represents the sequence of numbers.}
}
\value{
The list of data frames with year and month columns.
}
\description{
This function reads archives by specified years from working directory and
return a list of data frames with year and month columns
}
\examples{
fars_read_years(2019)
fars_read_years(c(2019, 2018, "2017"))

\dontrun{
fars_read_years("a character")
}

}
