% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Count number of entries by year and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A character or numeric vector which represents the sequence of numbers.}
}
\value{
a data frame with number of entries by year and month.
}
\description{
This function reads archives by specified years from working directory and
count number of entries in files by year and month
}
\examples{
fars_summarize_years(2019)
fars_summarize_years(c(2019, 2018, "2017"))

\dontrun{
fars_summarize_years("a character")
}

}
